pyproject.toml
[project]
name = "slp-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.4.2",
    "openai>=2.21.0",
    "pandas>=2.3.3",
    "plotly>=6.5.2",
    "pyarrow>=23.0.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "streamlit>=1.54.0",
]
